Add smart copy using date-diff

+++ compile.py
@@ -8,15 +8,19 @@ import argparse
 parser = argparse.ArgumentParser(description = 'Preprocess code against config file')
 parser.add_argument('-d', '--debug',
     help='Output code will have debug flag turned on', action='store_true')
-parser.add_argument('-r', '--rebuild',
-    help='Force new code generation, don\'t time-stamp check', action='store_true')
+parser.add_argument('-s', '--smart',
+    help='Use smart-update to only copy new files', action='store_true')
 parser.add_argument('-p', '--prefix', default="",
     help='Prefix of all output file names')
+
+parser.add_argument('--clean', help='Clean full output directory', action='store_true')
 
 parser.add_argument('--island', help='Only build missions for certain set of island(s)', action='append')
 parser.add_argument('--mod', help='Only build missions for certain set of mod(s)', action='append')
 
 parser.add_argument('--pbo', help='Build pbos', action='store_true')
+
+parser.add_argument('--verbose', '-v', help='Output additional log messages', action='store_true')
 
 args = parser.parse_args(sys.argv[1:])
 
@@ -29,10 +33,13 @@ islands = data['Islands'];
 missions = data['Missions'];
 cpbo = data['cpbo'];
 
+if args.clean:
+    shutil.rmtree(data['BuildDir'])
+
 if not os.path.isdir(data['BuildDir']):
     os.mkdir(data['BuildDir'])
 
-# Delete all base-files in the builddir (What are these? pbo?)
+# Delete all base-files in the builddir, mainly, the pbo's
 for the_file in os.listdir(data['BuildDir']):
     file_path = os.path.join(data['BuildDir'], the_file)
     if os.path.isfile(file_path):
@@ -45,11 +52,11 @@ for mission in missions:
 
     # Skip mods if we're only building a subset
     if args.island and not islandname in args.island:
-       print ('Skipping island ', islandname)
-       continue
+        if args.verbose: print ('Skipping island ', islandname)
+        continue
     if args.mod and not modname in args.mod:
-       print ('Skipping mod ', modname)
-       continue
+        if args.verbose: print ('Skipping mod ', modname)
+        continue
 
     print ('Creating a mission with mod', modname, 'on', islandname)
 
@@ -74,13 +81,32 @@ for mission in missions:
     missiondir = data['BuildDir'] + '/' + args.prefix + mission['name']+'.'+ missionIsland['class']
 
     # Clean up old files
-    if os.path.exists(missiondir):
-        shutil.rmtree(missiondir)
+    if not args.smart:
+        if os.path.exists(missiondir):
+            shutil.rmtree(missiondir)
+
+    def smartCopy(srcdir, dstdir):
+        if args.smart:
+            # If in smart mode, only copy files that have changed
+            for root, dirs, files in os.walk(srcdir):
+                for file in files:
+                    full_src = os.path.join(root, file)
+                    full_dst = os.path.join(dstdir, os.path.relpath(root, srcdir), file)
+
+                    # Only process file if it doesn't exist or is old
+                    if not os.path.isfile(full_dst) or int(os.path.getmtime(full_src)) > int(os.path.getmtime(full_dst)):
+                        if args.verbose: print "COPY %s TO %s" % (full_src, full_dst)
+                        try: os.makedirs(os.path.dirname(full_dst))
+                        except: pass
+                        shutil.copy2(full_src, full_dst)
+                        
+        else:
+            shutil.copytree(srcdir, dstdir)
 
     # Copy over code
-    shutil.copytree(data['Code']+'/',missiondir)
-    shutil.copytree(missionIsland['path']+'/',missiondir+'/Island/')
-    shutil.copytree(missionMod['path']+'/',missiondir+'/Units')
+    smartCopy(data['Code']+'/',missiondir)
+    smartCopy(missionIsland['path']+'/',missiondir+'/Island/')
+    smartCopy(missionMod['path']+'/',missiondir+'/Units')
     
     # Build require str
     required = ''
@@ -97,6 +123,7 @@ for mission in missions:
 
     # Pre-process code
     for rfile in data['ParsedFiles']:
+        if args.verbose: print "Processing file: %s..." % rfile
         full_rfile = os.path.join(missiondir, rfile)
         if not os.path.isfile(full_rfile):
             raise NameError('Unable to find file ' + rfile)
